# openapi.yaml
openapi: 3.1.0
info:
  title: Payment API
  description: RESTful API for secure payment processing
  contact:
    name: Payment Team
    email: payment@example.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local development
- url: https://api.payment.example.com
  description: Production
security:
- bearerAuth: []
...
paths:
...
  /api/v1/payments:
    get:
      tags:
      - Payments
      summary: Get payments (paginated + filtered)
      description: "All filter params are optional. Combine freely: ?status=FAILED&currency=USD&amountFrom=100&amountTo=500"
      operationId: getPayments
      parameters:
      - name: status
        in: query
        description: "Filter by status (PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED,\
          \ REVERSED, REFUNDED)"
        required: false
        schema:
          type: string
      - name: dateFrom
        in: query
        description: "Filter by creation date — from (ISO-8601, e.g. 2026-01-01T00:00:00)"
        required: false
        schema:
          type: string
          format: date-time
      - name: dateTo
        in: query
        description: "Filter by creation date — to (ISO-8601, e.g. 2026-12-31T23:59:59)"
        required: false
        schema:
          type: string
          format: date-time
      - name: amountFrom
        in: query
        description: Filter by minimum amount (inclusive)
        required: false
        schema:
          type: number
      - name: amountTo
        in: query
        description: Filter by maximum amount (inclusive)
        required: false
        schema:
          type: number
      - name: currency
        in: query
        description: "Filter by ISO 4217 currency code (e.g. USD, EUR, GBP)"
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: Payments retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelPaymentResponse"
        "400":
          description: Invalid filter parameters
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelPaymentResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelPaymentResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Payments
      summary: Create a new payment
      description: "Starts a durable Temporal workflow. Returns 202 Accepted + workflowId.\
        \ Poll GET /api/v1/payments/workflows/{workflowId}/status until a paymentId\
        \ is available, then GET /api/v1/payments/{paymentId}. Optionally supply an\
        \ 'Idempotency-Key' header to prevent duplicate charges."
      operationId: createPayment
      parameters:
      - name: Idempotency-Key
        in: header
        description: Client-generated unique key to prevent duplicate payments (UUID
          recommended)
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
        required: true
      responses:
        "202":
          description: Payment workflow started
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentWorkflowResponse"
        "400":
          description: Invalid payment request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentWorkflowResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentWorkflowResponse"
      security:
      - bearerAuth: []
  /api/v1/payments/{id}/reversal:
    post:
      tags:
      - Payments
      summary: Initiate a payment reversal or partial refund
      operationId: initiatePaymentReversal
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReversalRequest"
        required: true
      responses:
        "200":
          description: Reversal processed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Payment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "422":
          description: Reversal not possible for current payment state
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/payments/{id}/retry:
    post:
      tags:
      - Payments
      summary: Retry a failed payment
      description: Re-attempts a FAILED payment. Returns 409 if the payment is not
        in FAILED status.
      operationId: retryPayment
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Retry initiated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Payment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "409":
          description: Payment is not in FAILED status
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/payments/{id}/cancel:
    post:
      tags:
      - Payments
      summary: Cancel a payment
      description: Only PENDING payments can be cancelled. Returns 409 for any other
        status.
      operationId: cancelPayment
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Payment cancelled successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Payment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "409":
          description: Payment cannot be cancelled in its current status
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/payments/workflows/{workflowId}/cancel:
    post:
      tags:
      - Payments
      summary: Request graceful cancellation of a running payment workflow
      description: "Sends a cancellation signal to the workflow. It finishes the activity\
        \ currently in flight (to avoid a partial transfer) and then routes through\
        \ saga compensation on the next checkpoint, marking the payment FAILED with\
        \ reason 'cancelled-by-user'. Returns 202 once the signal is accepted — poll\
        \ GET /workflows/{workflowId}/status to observe CANCELLED."
      operationId: cancelWorkflow
      parameters:
      - name: workflowId
        in: path
        description: Workflow ID returned by POST /api/v1/payments
        required: true
        schema:
          type: string
      - name: reason
        in: query
        description: "Optional free-text reason for cancellation (audit trail, max\
          \ 500 chars)"
        required: false
        schema:
          type: string
          default: user-initiated
      responses:
        "202":
          description: Cancellation signal accepted
        "404":
          description: Workflow not found
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
...
  /api/v1/payments/{id}/status:
    patch:
      tags:
      - Payments
      summary: Update payment status
      operationId: updatePaymentStatus
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentStatusRequest"
        required: true
      responses:
        "200":
          description: Payment status updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Payment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "409":
          description: Invalid status transition
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
...
  /api/v1/payments/{id}:
    get:
      tags:
      - Payments
      summary: Get a payment by ID
      operationId: getPaymentById
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Payment found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Payment not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Payments
      summary: Delete a payment
      operationId: deletePayment
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Payment deleted successfully
        "404":
          description: Payment not found
        "409":
          description: Cannot delete a completed payment
        "500":
          description: Internal server error
      security:
      - bearerAuth: []
  /api/v1/payments/{id}/transactions:
    get:
      tags:
      - Payments
      summary: List transactions for a payment
      description: Returns all transaction records associated with the given payment.
        Non-admin users may only retrieve transactions for payments they own.
      operationId: getTransactionsByPaymentId
      parameters:
      - name: id
        in: path
        description: Payment UUID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Transactions retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "404":
          description: Payment not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "403":
          description: "Access denied: payment belongs to another user"
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    PaymentRequest:
      type: object
      description: Request payload for creating a new payment
      properties:
        sourceAccount:
          type: string
          description: 10-digit numeric source account number
          example: "1234567890"
          minLength: 1
          pattern: "\\d{10}"
        destinationAccount:
          type: string
          description: 10-digit numeric destination account number
          example: "0987654321"
          minLength: 1
          pattern: "\\d{10}"
        amount:
          type: number
          description: "Payment amount (positive, up to 4 decimal places, max 1,000,000)"
          example: 100.0
          maximum: 1000000.00
          minimum: 0.01
        currency:
          type: string
          description: ISO 4217 currency code
          example: USD
          minLength: 1
          pattern: "[A-Z]{3}"
        description:
          type: string
          description: "Optional payment reference or description (e.g. invoice number,\
            \ memo)"
          example: "Invoice #INV-2026-001"
          maxLength: 255
          minLength: 0
      required:
      - amount
      - currency
      - destinationAccount
      - sourceAccount
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PagedModelPaymentResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/PaymentResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT